refactor: move WithNodeRelease to providers
